+++ src/stores/useCount.js
@@ -1,0 +1,8 @@
+export default function useCount() {
+  function useMonthlyAmount(arr = []) {
+    const total = arr.reduce((s, v) => Number((s += v.amount)), 0)
+    return total.toLocaleString()
+  }
+
+  return { useMonthlyAmount }
+}
